refactor(core): use reasonably specific import paths
