+++ src/plugins/iconMenu/types.ts
@@ -1,5 +1,5 @@
 import type { Component } from 'vue'
-import type { PluginContainer, PluginOptions } from '@/core'
+import type { Icon, PluginContainer, PluginOptions } from '@/core'
 
 export const PluginId = 'iconMenu'
 
@@ -16,7 +16,7 @@ export interface Menu {
 	 * Current examples for the usage without icon include Zoom and Fullscreen if
 	 * {@link MapConfiguration.layout | `layout`} is set to `'nineRegions'`
 	 */
-	icon?: string
+	icon?: Icon
 }
 
 /**
